Make code more general; options; element spec
Start making code more general. I understand how to use BeautifulSoup to grab
the contents of a web page, so now it's time to build a program around it.

First of all, I added some program options using the argparse module provided by
Python. This is a real blessing and makes working with command line parameters
very easy.

Then I started the parsing of a specifier string for the interesting HTML
elements. For now, you can extract a string from the HTML by using the specifier
format <CSS selector>[/<attribute name>], which will select the element by its
CSS path/selector; if given, the content of the specified attribute is used; if
this is not given or empty, the tag's string content is used.