Use lxml directly instead of using BeautifulSoup
BeautifulSoup is a nice framework, but we only need a very small part of actual
HTML parsing and model access in this program. Since we decided to use CSS
selectors, lxml is actually a better choice for the HTML parser.
The decision to switch over from a working prototype using BeautifulSoup to lxml
is that many webcomics don't use id's and classes a lot, so selecting the
required elements from the HTML document sometimes takes all the power of CSS3
selectors. BeautifulSoup only implements a certain subset of those, most notably
lacking the + (next sibling) operator.