Add support for optionally using the requests lib
Since the builtin Python urllib2/httplib likes to barf on some stuff which might
be responded by httpds, you can use the requests library for those websites. If
you want to use requests, simply install it. The script will automatically use
it if it is available. Otherwise it will fallback to urllib2+httplib.