Fix file format to prevent regexp trouble
The file format of the templates and aliases files could previously contain
comments trailing a relevant line. This might have caused problems with regular
expressions containing a pipe character (which is used as an alternative in
regexp).
To prevent this, lines may now only consist of exactly two parts, divided by a
single pipe character (and some optional whitespace). In the templates file, the
split will be done left-oriented, in the aliases file right-oriented, and each
time with a maxsplit parameter of 1, so that any regexp will be left alone.